Separate mapBox actions from other redux actions
